add: arquivos do pull e organização das pastas do front
