+++ src/interface/views/alertas.py
@@ -1,0 +1,12 @@
+import streamlit as st
+
+
+def alertas_page():
+
+    # Header
+    st.markdown("""
+    <div class="header-ipea">
+        <h3 class="titulo-ipea">Relatórios inteligentes IPEA</h3>
+    </div>
+    """, unsafe_allow_html=True)
+    st.write("Conteúdo da página de Alertas")
