--- src/interface/views/avisos.py
@@ -1,5 +1,0 @@
-import streamlit as st
-
-def render():
-    st.markdown("# Dashboard IPEA")
-    # conteúdo da dashboard aqui
