+++ src/interface/views/configuracoes.py
@@ -1,0 +1,6 @@
+import streamlit as st
+
+
+def configuracoes_page():
+    st.title("Configurações")
+    st.write("Conteúdo da página de Configurações")
