+++ src/interface/views/user.py
@@ -1,0 +1,6 @@
+import streamlit as st
+
+
+def user_page():
+    st.title("Perfil do Usuário")
+    st.write("Conteúdo da página do Usuário")
